+++ app/Http/Controllers/ProductController.php
@@ -18,10 +18,12 @@ class ProductController extends Controller
     public function index(IndexProductRequest $request): AnonymousResourceCollection
     {
         $query = Product::with([
+            'brand',
             'productType',
             'barcodes',
             'currency',
             'productPresentations.presentation.presentationType',
+            'productPresentations.barcodes',
         ]);
 
         if ($request->filled('search')) {
@@ -57,7 +59,7 @@ class ProductController extends Controller
         $product->update(['is_active' => ! $product->is_active]);
 
         return ProductResource::make(
-            $product->fresh()->load(['productType', 'barcodes', 'currency', 'productPresentations.presentation.presentationType'])
+            $product->fresh()->load(['brand', 'productType', 'barcodes', 'currency', 'productPresentations.presentation.presentationType', 'productPresentations.barcodes'])
         );
     }
 }