+++ app/Models/Product.php
@@ -7,6 +7,7 @@ use App\Models\Concerns\HasUuid;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\HasManyThrough;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Product extends Model
@@ -24,6 +25,11 @@ class Product extends Model
         ];
     }
 
+    public function brand(): BelongsTo
+    {
+        return $this->belongsTo(Brand::class);
+    }
+
     public function productType(): BelongsTo
     {
         return $this->belongsTo(ProductType::class);
@@ -34,9 +40,9 @@ class Product extends Model
         return $this->belongsTo(Currency::class);
     }
 
-    public function barcodes(): HasMany
+    public function barcodes(): HasManyThrough
     {
-        return $this->hasMany(Barcode::class);
+        return $this->hasManyThrough(Barcode::class, ProductPresentation::class);
     }
 
     public function productPresentations(): HasMany